refactor of package_tests
